+++ NEWS.md
@@ -13,6 +13,8 @@ Bug fixes:
 + Fix false negatives in PHP5 backwards compatibility heuristic checks (#1939)
 + Fix false positive PhanUnanalyzableInheritance for a method inherited from a trait (which itself uses trait) (#1968)
 + Fix an uncaught RuntimeException when type checking an array that was roughly 12 or more levels deep (#1962)
++ Improve checks of the return type of magic methods against methods inherited from ancestor classes (#1975)
+  Don't emit a false positive `PhanParamSignaturePHPDocMismatchReturnType`
 
 Language Server/Daemon mode:
 + Fix an uncaught exception when extracting a URL with an unexpected scheme (not `file:/...`) (#1960)